PoW - refactoring: ethpow moved to its own package
